sB/requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
redis==4.0.2
prometheus-client==0.11.0
grpcio==1.41.0
protobuf==3.19.0
websockets==10.0
python-dotenv==0.19.0